+++ defaultModel/src/main/scala/pl/touk/nussknacker/defaultmodel/DefaultModelMigrations.scala
@@ -14,6 +14,7 @@ class DefaultModelMigrations extends ProcessMigrations {
     // 100 -> NewMigration,
     // Newly added migrations should be in the hundreds: 100, 200, 300 and so on. We do this because
     // many ProcessMigrations can be loaded using SPI, and we want to avoid overlapping numbers when merging.
+    100 -> SampleGeneratorToEventGeneratorAndPeriodToScheduleParameter
   )
 
 }